requirements.txt
google-generativeai
openai==0.28
python-dotenv
streamlit